cpu_server/requirements.txt
transformers
torch
pydantic
fastapi[standard]
Pillow